Struct netflowv9::Header

pub struct Header<'bytes> {
    // some fields omitted
}

Methods

impl<'bytes> Header<'bytes>

fn version(&self) -> u16

Version of Flow Record format exported in this packet. The value of this field is 9 for the current version.

fn count(&self) -> u16

The total number of records in the Export Packet, which is the sum of Options FlowSet records, Template FlowSet records, and Data FlowSet records.

fn sys_uptime(&self) -> u32

Time in milliseconds since this device was first booted.

fn seconds(&self) -> u32

Time in seconds since 0000 UTC 1970, at which the Export Packet leaves the Exporter.

fn seq_num(&self) -> u32

Incremental sequence counter of all Export Packets sent from the current Observation Domain by the Exporter. This value MUST be cumulative, and SHOULD be used by the Collector to identify whether any Export Packets have been missed.

fn source_id(&self) -> u32

A 32-bit value that identifies the Exporter Observation Domain. NetFlow Collectors SHOULD use the combination of the source IP address and the Source ID field to separate different export streams originating from the same Exporter.